More stuff before merge

+++ ArduPlane/RALPHIE_UNIT_TEST.h
@@ -1,5 +1,5 @@
 class RALPHIEUnitTest {
 
     public:
-        float testUNIT(float a, float b);
+        int testUNIT(int a, int b);
 };
